debian: add xcb flavour
This commit eliminates the previous hack with the onefile. It was
easier to maintain in theory, but in practice not.

With the introduction of the xcb flavour, the annotation were
removed - these was options which was enforced by the ubuntu
maintainers during build kernel.

In case of kernel config conflict, you should run
```
debian/rules updateconfigs
```
to properly update the common kernel options.

Bla bla bla, ha kell meg ide info, akkor tudok irni...

https://wiki.ubuntu.com/Kernel/BuildYourOwnKernel
https://help.ubuntu.com/community/Kernel/Compile